+++ src/components/layout/cart/CartTotal.tsx
@@ -1,122 +1,16 @@
-/* eslint-disable react/button-has-type */
 import { css } from '@emotion/react';
-import { useRecoilState, useSetRecoilState } from 'recoil';
-import { useNavigate } from 'react-router-dom';
-import swal from 'sweetalert';
-import { cartPostToJudgment } from '../../../api/cart';
-import {
-  cartDataState,
-  cartSoldOutState,
-  cartCheckedRoomListState,
-} from '../../../atoms/cartAtom';
+import { useRecoilValue } from 'recoil';
+import { cartCheckedRoomListState } from '../../../atoms/cartAtom';
 import theme from '../../../style/theme';
+import CartOrder from './CartOrder';
 
 interface CartTotalProps {
   totalPrice: number;
 }
 
 const CartTotal = ({ totalPrice }: CartTotalProps) => {
-  const navigate = useNavigate();
-  const setCartData = useSetRecoilState(cartDataState);
-  const [soldOutData, setSoldOutData] = useRecoilState(cartSoldOutState);
-  const [checkedRoomList, setCheckedRoomList] = useRecoilState(
-    cartCheckedRoomListState,
-  );
+  const checkedRoomList = useRecoilValue(cartCheckedRoomListState);
 
-  const handlePostClick = async () => {
-    const updateCheckboxAvailability = () => {
-      const unavailableRoomIds = soldOutData?.unavailableIds || [];
-      const updatedCheckedRoomList =
-        checkedRoomList.map((room) => ({
-          ...room,
-          disabled: unavailableRoomIds.includes(room.roomId),
-        })) || [];
-      setCheckedRoomList(updatedCheckedRoomList);
-    };
-    try {
-      if (checkedRoomList.length > 3) {
-        swal({
-          title: '예약 한도 초과',
-          text: '3개 이상의 상품을 주문할 수 없습니다',
-          icon: 'error',
-        });
-        return;
-      }
-
-      if (checkedRoomList.length === 0) {
-        swal({
-          title: '선택한 상품이 없습니다.',
-          icon: 'error',
-        });
-        return;
-      }
-
-      const rooms = checkedRoomList.map(({ roomId, startDate, endDate }) => ({
-        roomId,
-        startDate,
-        endDate,
-      }));
-
-      const updatedCartData = checkedRoomList.map(
-        ({
-          roomId,
-          roomName,
-          startDate,
-          endDate,
-          productName,
-          standard,
-          capacity,
-          checkIn,
-          checkOut,
-          price,
-        }) => ({
-          roomId,
-          roomName,
-          productName,
-          startDate,
-          endDate,
-          standard,
-          capacity,
-          checkIn,
-          checkOut,
-          price,
-        }),
-      );
-
-      const response = await cartPostToJudgment(rooms);
-
-      if (!response) {
-        swal({
-          title: '주문 가능',
-          text: '주문 가능한 상품입니다.',
-          icon: 'success',
-        });
-        navigate('/pay');
-        setCartData(updatedCartData);
-        setCheckedRoomList([]);
-      } else {
-        setSoldOutData(response);
-        updateCheckboxAvailability();
-
-        const soldOutRoomNames = checkedRoomList
-          .filter(
-            (room) =>
-              response.unavailableIds &&
-              response.unavailableIds.includes(room.roomId),
-          )
-          .map((room) => room.productName)
-          .join(', ');
-        swal({
-          title: '상품 품절',
-          text: `[${soldOutRoomNames}] 상품이 품절되었습니다. 주문할 수 없습니다.`,
-          icon: 'error',
-        });
-        setCheckedRoomList([]);
-      }
-    } catch (error) {
-      console.error(error);
-    }
-  };
   return (
     <div css={Container}>
       <div css={InnerContainer}>
@@ -133,13 +27,7 @@ const CartTotal = ({ totalPrice }: CartTotalProps) => {
         ) : (
           ''
         )}
-        <button
-          onClick={handlePostClick}
-          css={checkedRoomList.length > 3 ? DisabledButton : OrderButton}
-          disabled={checkedRoomList.length > 3}
-        >
-          주문하기
-        </button>
+        <CartOrder />
       </div>
     </div>
   );
@@ -180,39 +68,4 @@ const ErrorText = css`
   color: ${theme.colors.error};
   font-size: 1rem;
   font-weight: 700;
-`;
-
-const OrderButton = css`
-  width: 70%;
-
-  margin: auto;
-  margin-top: 1rem;
-  padding: 1rem;
-
-  border-radius: 0.5rem;
-
-  font-size: 1.3rem;
-  font-weight: 700;
-  color: ${theme.colors.white};
-  background-color: ${theme.colors.blue700};
-
-  cursor: pointer;
-  transition: all 0.15s ease-in-out;
-
-  &:hover {
-    background-color: ${theme.colors.blue600};
-  }
-`;
-
-const DisabledButton = css`
-  width: 70%;
-  margin: auto;
-  margin-top: 1rem;
-  padding: 1rem;
-  border-radius: 0.5rem;
-  font-size: 1.3rem;
-  font-weight: 700;
-  color: ${theme.colors.white};
-  background-color: ${theme.colors.gray400};
-  cursor: not-allowed;
 `;
